backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.19
supabase>=2.0.0
python-dotenv>=1.0.0
transformers>=4.30.0
torch>=2.0.0
requests>=2.30.0
pydantic-settings>=2.0.0
numpy>=1.24.0
PyJWT>=2.8.0
